fixing add redirect uri to management
